CreateProcessW -> ShellExecuteW
For whatever absolutely-cursed-for-no-goddamn-reason reason, launching
libqt executables with CreateProcess causes QFileDialog to show the qt
version instead of the native version. WHO THE HELL IS RESPONSIBLE FOR
THIS ABSOLUTE GARBAGE? Anyways, there was one last trick up my sleeve
and that is ShellExecute. Is there a good reason for why? No. Does it
do what I want? Yes, moderately. It's safe to say I'm going to stay
far away from this lib in the future.